+++ packages/yourloops/lib/team/team.util.ts
@@ -74,7 +74,7 @@ export default class TeamUtils {
   }
 
   static isPrivate(team: Team): boolean {
-    return team.id === PRIVATE_TEAM_ID
+    return team?.id === PRIVATE_TEAM_ID
   }
 
   static formatTeamNameForTestId(teamName: string): string {